Rename: 패키지명 변경 해결 (#56)
