+++ src/main/java/com/adoonge/seedzip/recommendation/service/YouTubeService.java
@@ -1,4 +1,4 @@
-package com.adoonge.seedzip.summary.service;
+package com.adoonge.seedzip.recommendation.service;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
